Changed the data type used by the factorial function

+++ lib/math/math.hpp
@@ -8,7 +8,7 @@ namespace pix::math
      * @param num Number
      * @return Factorial value
     */
-    unsigned int fat(unsigned int);
+    unsigned long fat(unsigned long);
 }
 
 // Implementation file
